feat: separate logic from register component into custom hook
