+++ src/components/Register/Register.tsx
@@ -1,40 +1,20 @@
-import { useState } from 'react'
-import { useNavigate } from 'react-router-dom'
-import { useAddUserMutation } from '../../store/slices/usersApiSlice'
+import useRegister from '../../hooks/useRegister'
 import { RegisterContainer, ButtonContainer, RegisterButton, LoginButton } from './styles'
 
 const Register = () => {
-  const navigate = useNavigate()
-  const [createUser] = useAddUserMutation()
-
-  const [first_name, setFirstName] = useState('')
-  const [last_name, setLastName] = useState('')
-  const [email, setEmail] = useState('')
-  const [password1, setPassword1] = useState('')
-  const [password2, setPassword2] = useState('')
-
-  const handleRegister = (e: React.SyntheticEvent) => {
-    e.preventDefault()
-
-    if (first_name === '') {
-      return alert('First name cannot be blank. Please try again.')
-    }
-    if (last_name === '') {
-      return alert('Last name cannot be blank. Please try again.')
-    }
-    if (email === '') {
-      return alert('Email cannot be blank. Please try again.')
-    }
-    if (password1 === '' || password2 === '') {
-      return alert('Password cannot be blank. Please try again.')
-    }
-    if (password1 !== password2) {
-      return alert('Passwords do not match. Please try again.')
-    }
-
-    createUser({first_name, last_name, email, password1})
-    navigate('/Login')
-  }
+  const {
+    first_name,
+    setFirstName,
+    last_name,
+    setLastName,
+    email,
+    setEmail,
+    password1,
+    setPassword1,
+    password2,
+    setPassword2,
+    handleRegister,
+  } = useRegister()
 
   return (
     <>
@@ -43,30 +23,35 @@ const Register = () => {
           type="text"
           placeholder="Enter first name"
           name="first name"
+          value={first_name}
           onChange={(e) => setFirstName(e.target.value)}
         />
         <input
           type="text"
           placeholder="Enter last name"
           name="last name"
+          value={last_name}
           onChange={(e) => setLastName(e.target.value)}
         />
         <input
           type="text"
           placeholder="Enter email"
           name="email"
+          value={email}
           onChange={(e) => setEmail(e.target.value)}
         />
         <input
           type="password"
           placeholder="Enter password"
           name="password"
+          value={password1}
           onChange={(e) => setPassword1(e.target.value)}
         />
         <input
           type="password"
           placeholder="Enter password again"
           name="passwordCheck"
+          value={password2}
           onChange={(e) => setPassword2(e.target.value)}
         />
       </RegisterContainer>
